Tentativa de Update

+++ DAM_TP1/src/main/kotlin/dam/exer_v1/PhysicalBook.kt
@@ -5,7 +5,7 @@ class PhysicalBook(
     author: String,
     publicationYear: Int,
     availableCopies: Int,
-    var weight: Double,
+    var weight: Int,
     var hasHardover: Boolean = true,
 ) : Book(title, author, publicationYear, availableCopies)
 {
